+++ test/unit/array/array-monotonic.predicate.spec.ts
@@ -32,6 +32,10 @@ describe('arrayMonotonic', () => {
         it('when compared using a comparator', () => {
             expect(arrayMonotonic(values, { comparator: direct_comparator, monotonicity })).toBeTrue()
         })
+
+        it('when compared ordinally', () => {
+            expect(arrayMonotonic(values, { monotonicity })).toBeTrue()
+        })
     })
 
     describe.each<[number[], Monotonicity]>([
@@ -51,6 +55,10 @@ describe('arrayMonotonic', () => {
 
         it('when compared using a comparator', () => {
             expect(arrayMonotonic(values, { comparator: direct_comparator, monotonicity })).toBeFalse()
+        })
+
+        it('when compared ordinally', () => {
+            expect(arrayMonotonic(values, { monotonicity })).toBeFalse()
         })
     })